Task.1 Commit

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,5 @@
 #include "function_pointers.h"
+#include "stdlib.h"
 /**
  * array_iterator - a function as parameter
  *                  on each element of an array.
@@ -17,7 +18,7 @@ void array_iterator(int *array, size_t size, void (*action)(int))
 
 	if (array == NULL  && action = NULL)
 	{
-		for (i = 0 ; i  < size ; i++)
+		for (i = 0 ; i  < (int)  size ; i++)
 		{
 			action(array[i]);
 		}
